Update documentation setup getting error
"To simplify URL configs, patterns() was deprecated in Django 1.8, and removed in 1.10 (release notes). In Django 1.10, urlpatterns must be a list of url() instances. Using a tuple in patterns() is not supported any more, and the Django checks framework will raise an error."
source: https://stackoverflow.com/questions/38786461/django-error-your-url-pattern-is-invalid-ensure-that-urlpatterns-is-a-list-of

Error log:
?: (urls.E004) Your URL pattern ('^favicon\\.ico$', <function RedirectView at 0x11283a9d8>) is invalid. Ensure that urlpatterns is a list of path() and/or re_path() instances.
        HINT: Try using path() instead of a tuple.