Fix entry header image issue in blog page
### Issue summary
When an entry does not have a header image it might display the header image of another entry in the list of entries.

### How to reproduce?
Create an entry without a header image (entry 1), then create a second entry  (entry 2) with a header image. visit blog page, entry 1 will appear with the entry 2 image.